Reducing size of minimal component
