+++ src/motion/utils/use-visual-element.ts
@@ -1,4 +1,4 @@
-import { useContext, useEffect } from "react"
+import { MutableRefObject, useContext, useEffect, useRef } from "react"
 import { PresenceContext } from "../../context/PresenceContext"
 import { isPresent } from "../../components/AnimatePresence/use-presence"
 import { LayoutGroupContext } from "../../context/LayoutGroupContext"
@@ -6,7 +6,6 @@ import { MotionProps } from "../../motion"
 import { useVisualElementContext } from "../../context/MotionContext"
 import { useSnapshotOnUnmount } from "../../motion/features/layout/use-snapshot-on-unmount"
 import { CreateVisualElement, VisualElement } from "../../render/types"
-import { useConstant } from "../../utils/use-constant"
 import { useIsomorphicLayoutEffect } from "../../utils/use-isomorphic-effect"
 import { MotionConfigContext } from "../../context/MotionConfigContext"
 import { VisualState } from "./use-visual-state"
@@ -19,27 +18,38 @@ function useLayoutId({ layoutId }: MotionProps) {
 }
 
 export function useVisualElement<Instance, RenderState>(
-    isStatic: boolean,
     visualState: VisualState<Instance, RenderState>,
-    createVisualElement: CreateVisualElement<Instance>,
-    props: MotionProps
-): VisualElement<Instance> {
+    props: MotionProps,
+    createVisualElement?: CreateVisualElement<Instance>
+): VisualElement<Instance> | undefined {
     const config = useContext(MotionConfigContext)
     const parent = useVisualElementContext()
     const presenceContext = useContext(PresenceContext)
     const layoutId = useLayoutId(props)
 
-    const visualElement = useConstant(() =>
-        createVisualElement(isStatic, {
+    const visualElementRef: MutableRefObject<
+        VisualElement | undefined
+    > = useRef(undefined)
+
+    if (createVisualElement === null) {
+        createVisualElement = config.visualElement
+    }
+
+    if (!visualElementRef.current && createVisualElement) {
+        visualElementRef.current = createVisualElement({
             visualState,
             parent,
             props: { ...props, layoutId },
             presenceId: presenceContext?.id,
             blockInitialAnimation: presenceContext?.initial === false,
         })
-    )
+    }
+
+    const visualElement = visualElementRef.current
 
     useIsomorphicLayoutEffect(() => {
+        if (!visualElement) return
+
         visualElement.setProps({
             ...config,
             ...props,
@@ -57,6 +67,8 @@ export function useVisualElement<Instance, RenderState>(
     })
 
     useEffect(() => {
+        if (!visualElement) return
+
         /**
          * In a future refactor we can replace the features-as-components and
          * have this loop through them all firing "effect" listeners
