+++ src/render/dom/utils/create-config.ts
@@ -6,11 +6,13 @@ import { HTMLRenderState } from "../../html/types"
 import { SVGRenderState } from "../../svg/types"
 import { svgMotionConfig } from "../../svg/config-motion"
 import { htmlMotionConfig } from "../../html/config-motion"
+import { CreateVisualElement } from "../../types"
 
 export function createDomMotionConfig<Props>(
     defaultFeatures: MotionFeature[],
     Component: string | React.ComponentType<Props>,
-    forwardMotionProps: boolean
+    forwardMotionProps: boolean,
+    createVisualElement?: CreateVisualElement<any>
 ) {
     const baseConfig = isSVGComponent(Component)
         ? svgMotionConfig
@@ -20,6 +22,7 @@ export function createDomMotionConfig<Props>(
         ...baseConfig,
         defaultFeatures,
         useRender: createUseRender(Component, forwardMotionProps),
+        createVisualElement,
     } as
         | MotionComponentConfig<SVGElement, SVGRenderState>
         | MotionComponentConfig<HTMLElement, HTMLRenderState>
