+++ src/render/html/config-motion.ts
@@ -3,14 +3,13 @@ import { makeUseVisualState } from "../../motion/utils/use-visual-state"
 import { HTMLRenderState } from "./types"
 import { scrapeMotionValuesFromProps as scrapeHTMLProps } from "./utils/scrape-motion-values"
 import { createHtmlRenderState } from "./utils/create-render-state"
-import { htmlVisualElement } from "./visual-element"
 
 export const htmlMotionConfig: Partial<MotionComponentConfig<
     HTMLElement,
     HTMLRenderState
 >> = {
-    createVisualElement: (isStatic, options) =>
-        htmlVisualElement(options, { enableHardwareAcceleration: !isStatic }),
+    // createVisualElement: (options) =>
+    //     htmlVisualElement(options, { enableHardwareAcceleration: true }),
     useVisualState: makeUseVisualState({
         scrapeMotionValuesFromProps: scrapeHTMLProps,
         createRenderState: createHtmlRenderState,
