+++ src/render/svg/utils/render.ts
@@ -1,0 +1,15 @@
+import { camelToDash } from "../../dom/utils/camel-to-dash"
+import { renderHTML } from "../../html/utils/render"
+import { SVGRenderState } from "../types"
+import { camelCaseAttributes } from "./camel-case-attrs"
+
+export function renderSVG(element: SVGElement, renderState: SVGRenderState) {
+    renderHTML(element as any, renderState)
+
+    for (const key in renderState.attrs) {
+        element.setAttribute(
+            !camelCaseAttributes.has(key) ? camelToDash(key) : key,
+            renderState.attrs[key] as string
+        )
+    }
+}
