+++ src/value/utils/is-motion-value.ts
@@ -1,5 +1,5 @@
 import { MotionValue } from "../"
 
 export const isMotionValue = (value: any): value is MotionValue => {
-    return value instanceof MotionValue
+    return Object.prototype.hasOwnProperty.call(value, "getVelocity")
 }
