+++ src/value/utils/resolve-motion-value.ts
@@ -1,6 +1,7 @@
 import { MotionValue } from ".."
 import { isCustomValue } from "../../utils/resolve-value"
 import { CustomValueType } from "../../types"
+import { isMotionValue } from "./is-motion-value"
 
 /**
  * If the provided value is a MotionValue, this returns the actual value, otherwise just the value itself
@@ -12,7 +13,7 @@ import { CustomValueType } from "../../types"
 export function resolveMotionValue(
     value?: string | number | CustomValueType | MotionValue
 ): string | number {
-    const unwrappedValue = value instanceof MotionValue ? value.get() : value
+    const unwrappedValue = isMotionValue(value) ? value.get() : value
     return isCustomValue(unwrappedValue)
         ? unwrappedValue.toValue()
         : unwrappedValue
